server/requirements.txt
accounts==0.2.4
anaconda-navigator==2.0.3
anaconda-project==0.9.1
argh==0.26.2
asn1crypto==1.4.0
astropy==4.2.1
autopep8==1.5.6
backports.shutil-get-terminal-size==1.0.0
bcrypt==4.3.0
bitarray==1.9.2
bkcharts==0.2
bokeh==2.3.2
boto==2.49.0
bottleneck==1.3.2
brotlipy==0.7.0
colorama==0.4.4
conda-build==3.21.4
conda-package-handling==1.7.3
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cython==0.29.23
cytoolz==0.11.0
distributed==2021.4.0
django-cors-headers==4.4.0
django-rest-framework==0.1.0
django-wsgi==1.0b1
djangorestframework-simplejwt==5.3.1
fastcache==1.1.0
flake8==3.9.0
flask==1.1.2
gevent==21.1.2
glob2==0.7
gmpy2==2.0.8
html5lib==1.1
iniconfig==1.1.1
jdcal==1.4.1
jupyter==1.0.0
jupyterlab==3.0.14
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
mock==4.0.3
mpi4py==3.0.3
multipledispatch==0.6.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
navigator-updater==0.2.1
neuron==8.0.0
nltk==3.6.1
nose==1.3.7
numba==0.53.1
olefile==0.46
openpyxl==3.0.7
pactools==0.3.1
path==15.1.2
pathlib2==2.3.5
pep8==1.7.1
pip-chill==1.0.3
ply==3.11
pooch==1.6.0
pycurl==7.43.0.6
pydocstyle==6.0.0
pymysql==1.1.1
pyodbc==4.0.0-unsupported
pyopenssl==20.0.1
pysocks==1.7.1
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
rope==0.18.0
rtree==0.9.7
scikit-image==0.18.1
seaborn==0.11.1
simplegeneric==0.8.1
singledispatch==0.0.0
sortedcollections==2.1.0
sphinxcontrib-websupport==1.2.4
spyder==4.2.5
statsmodels==0.12.2
sympy==1.8
tables==3.6.1
unicodecsv==0.14.1
xlrd==2.0.1
xlsxwriter==1.3.8
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.31.0
